+++ src/utils/learnUtil.ts
@@ -1,0 +1,8 @@
+export const shuffleArray = (array: Array<any>) => {
+  const shuffled = [...array];
+  for (let i = shuffled.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]];
+  }
+  return shuffled;
+};
